# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SMTP Sandbox API Reference
  version: 1.0.0
  description: |
    API REST para a plataforma de Sandbox SMTP e Webmail em Tempo Real.
    Permite criação de domínios, caixas postais, consulta de mensagens recebidas, download de anexos, disparo de webhooks e geração de API Keys.
servers:
  - url: /api
    description: Servidor de Desenvolvimento / Produção
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: Use sua chave no formato 'Bearer sk_live_...'
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /messages/{id}/raw:
    get:
      summary: Obter mensagem RAW RFC822
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Mensagem em texto bruto
          content:
            text/plain:
              schema: { type: string }
...
  /attachments/{id}:
    get:
      summary: Download de arquivo anexo
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Stream binário do arquivo
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
...